+++ packages/react-native-reanimated/src/common/style/registry.ts
@@ -9,6 +9,7 @@ import {
 } from './propsBuilder';
 
 const DEFAULT_SEPARATELY_INTERPOLATED_NESTED_PROPERTIES = new Set<string>([
+  'backgroundImage',
   'boxShadow',
   'shadowOffset',
   'textShadowOffset',