+++ packages/react-native-reanimated/src/css/native/__tests__/registry.test.ts
@@ -155,6 +155,7 @@ describe('registry', () => {
       );
       expect(props).toEqual(
         new Set([
+          'backgroundImage',
           'boxShadow',
           'shadowOffset',
           'textShadowOffset',